Testbed probe for Android paths.
